requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
defusedxml==0.7.1
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-cors-headers==3.14.0
django-phone-field==1.8.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
drfpasswordless==1.5.8
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
sendgrid==6.9.7
six==1.16.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
twilio==7.16.5
tzdata==2022.7
urllib3==1.26.14